Library page reads list of Users from DB
So, it turns out that database accesses must be done in alternate
threads, but UI updates can only be done in the main thread. To manage
this awful bout of communication, you need to create a Handler. This is
an implementation that reads the list fo Users from the database to the
Library page, but should be expandable to do everything we need.

I hate dealing with threads.